requirements.txt
asgiref==3.4.1
Django==3.2.25
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==21.3
pyparsing==3.0.7
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.1.1
uritemplate==4.1.1
